Menu Suggestions
The initial page caught my eye, but after examining other pages, I personally like the softness of the menu.  Looking at the site the third time the menu color was very hard and solid and the logo was hard and solid as well.  This document makes some suggestions of updating the logo and menu.